fix(contracts): solve formatting / linting issues

+++ benchmark/EcGasHelper.sol
@@ -40,11 +40,11 @@ contract EcGasHelper {
     }
 
     function _isOnCurve(
-        uint _x,
-        uint _y,
-        uint _aa,
-        uint _bb,
-        uint _pp
+        uint256 _x,
+        uint256 _y,
+        uint256 _aa,
+        uint256 _bb,
+        uint256 _pp
     ) external returns (bool) {
         return EllipticCurve.isOnCurve(_x, _y, _aa, _bb, _pp);
     }
