+++ contracts/helpers/TestEllipticCurve.sol
@@ -42,11 +42,11 @@ contract TestEllipticCurve {
     }
 
     function isOnCurve(
-        uint _x,
-        uint _y,
-        uint _aa,
-        uint _bb,
-        uint _pp
+        uint256 _x,
+        uint256 _y,
+        uint256 _aa,
+        uint256 _bb,
+        uint256 _pp
     ) public pure returns (bool) {
         return EllipticCurve.isOnCurve(_x, _y, _aa, _bb, _pp);
     }
